moving migrate-ns ast-util to a common file
